# ChallengeTwoApi/Models/ViewModels/Login.cs
namespace ChallengeTwoApi.Models.ViewModels
{
    public class Login
    {
        public string Email { get; set; } = string.Empty;

        public string Password { get; set; } = string.Empty;
    }
}