# ChallengeTwoApi/Models/ViewModels/Register.cs
namespace ChallengeTwoApi.Models.ViewModels
{
    public class Register
    {
        public string Email { get; set; } = string.Empty;

        public string FirstName { get; set; } = string.Empty;

        public string LastName { get; set; } = string.Empty;

        public string Password { get; set; } = string.Empty;
    }
}